requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9
defusedxml==0.6.0
Django==3.0.5
django-crispy-forms==1.9.0
django-friendship==1.9.1
django-hashid-field==3.1.1
hashids==1.2.0
idna==2.9
oauthlib==3.1.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.9
